done with everything except the new functionality

+++ Client Side Rendering Lab/cookBook/src/delete.js
@@ -1,17 +1,19 @@
+import {html, render} from '../node_modules/lit-html/lit-html.js';
+
+
 export function deleteRecipe(event) {
     const recipeId = event.target.parentNode.querySelector("p").textContent;
     const url = `http://localhost:3030/data/recipes/${recipeId}`;
     const token = JSON.parse(localStorage.getItem("authToken"));
 
-    fetch(url, {method: "DELETE", headers: {"X-Authorization": token}})
-    .then(response => {
-        if (response.status === 200 || response.status === 204) {
-            document.querySelector("article").remove();
-            document.querySelector("main").innerHTML += '<p style="color: whitesmoke;">Successful Deletion</p>';
-       
-        } else {
-            throw new Error("An error occurred");
-        }
-    })
-    .catch(err => console.log(err.message));
+    if (token) {
+        fetch(url, {method: "DELETE", headers: {"X-Authorization": token}})
+        .then(response => render(
+            html`<p style="color: whitesmoke;">Successful Deletion</p>`, 
+            document.querySelector("main")))
+        .catch(err => console.log(err.message));
+    
+    } else {
+        alert("You are not authenticated!");
+    }
 }
